+++ initialization/layout.go
@@ -2,7 +2,6 @@ package initialization
 
 import (
 	"fmt"
-	"math"
 
 	"github.com/spacemeshos/post/config"
 )
@@ -16,10 +15,9 @@ type filesLayout struct {
 
 func deriveFilesLayout(cfg config.Config, opts config.InitOpts) (filesLayout, error) {
 	maxFileNumLabels := opts.MaxFileNumLabels()
-	totalLabels := uint64(opts.NumUnits) * uint64(cfg.LabelsPerUnit)
 
 	firstFileIdx := opts.FromFileIdx
-	lastFileIdx := TotalFiles(cfg, opts) - 1
+	lastFileIdx := opts.TotalFiles(cfg.LabelsPerUnit) - 1
 
 	if opts.ToFileIdx != nil {
 		if *opts.ToFileIdx < 0 {
@@ -31,14 +29,14 @@ func deriveFilesLayout(cfg config.Config, opts config.InitOpts) (filesLayout, er
 		lastFileIdx = *opts.ToFileIdx
 	}
 
-	lastFileNumLabels := maxFileNumLabels
-	labelsLeft := totalLabels - firstLabelInFile(lastFileIdx, opts)
-	if labelsLeft < maxFileNumLabels {
-		lastFileNumLabels = labelsLeft
-	}
-
 	if firstFileIdx > lastFileIdx {
 		return filesLayout{}, fmt.Errorf("invalid range: last file index (%v) must be greater then first (%v)", lastFileIdx, firstFileIdx)
+	}
+
+	lastFileNumLabels := maxFileNumLabels
+	labelsLeft := opts.TotalLabels(cfg.LabelsPerUnit) - firstLabelInFile(lastFileIdx, opts)
+	if labelsLeft < maxFileNumLabels {
+		lastFileNumLabels = labelsLeft
 	}
 
 	numFiles := lastFileIdx - firstFileIdx + 1
@@ -53,9 +51,4 @@ func deriveFilesLayout(cfg config.Config, opts config.InitOpts) (filesLayout, er
 
 func firstLabelInFile(fileIdx int, opts config.InitOpts) uint64 {
 	return uint64(fileIdx) * opts.MaxFileNumLabels()
-}
-
-func TotalFiles(cfg config.Config, opts config.InitOpts) int {
-	totalLabels := uint64(opts.NumUnits) * uint64(cfg.LabelsPerUnit)
-	return int(math.Ceil(float64(totalLabels) / float64(opts.MaxFileNumLabels())))
 }
